requirements.txt
python-telegram-bot[job-queue]
pybit
python-dotenv
